requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic-settings>=2.1
firebase-admin>=6.4
google-auth>=2.29
httpx>=0.27
